Move meta to meta package and remove mappings.
