+++ modules/core/src/main/scala/doobie/util/meta/meta.scala
@@ -2,15 +2,17 @@
 // This software is licensed under the MIT License (MIT).
 // For more information see LICENSE or https://opensource.org/licenses/MIT
 
-package doobie.util
+package doobie.util.meta
+
+import java.sql.{PreparedStatement, ResultSet}
 
 import cats.Invariant
 import cats.data.NonEmptyList
 import doobie.enum.JdbcType
-import java.sql.{ PreparedStatement, ResultSet }
+import doobie.util.{Get, Put}
+
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.TypeTag
-import Predef._
 
 /**
  * Convenience for introducing a symmetric `Get`/`Put` pair into implicit scope, and for deriving
@@ -136,7 +138,7 @@ trait MetaConstructors {
 }
 
 trait MetaInstances { this: MetaConstructors =>
-  import doobie.enum.JdbcType.{ Boolean => JdbcBoolean, _ }
+  import doobie.enum.JdbcType.{Boolean => JdbcBoolean, _}
 
   /** @group Instances */
   implicit val GetPutInvariant: Invariant[Meta] =
@@ -229,82 +231,15 @@ trait MetaInstances { this: MetaConstructors =>
       List(LongVarBinary),
       _.getBytes(_), _.setBytes(_, _), _.updateBytes(_, _))
 
-  /** @group Instances */
-  implicit val DateMeta: Meta[java.sql.Date] =
-    Basic.one[java.sql.Date](
-      Date,
-      List(Char, VarChar, LongVarChar, Timestamp),
-      _.getDate(_), _.setDate(_, _), _.updateDate(_, _))
-
-  /** @group Instances */
-  implicit val TimeMeta: Meta[java.sql.Time] =
-    Basic.one[java.sql.Time](
-      Time,
-      List(Char, VarChar, LongVarChar, Timestamp),
-      _.getTime(_), _.setTime(_, _), _.updateTime(_, _))
-
-  /** @group Instances */
-  implicit val TimestampMeta: Meta[java.sql.Timestamp] =
-    Basic.one[java.sql.Timestamp](
-      Timestamp,
-      List(Char, VarChar, LongVarChar, Date, Time),
-      _.getTimestamp(_), _.setTimestamp(_, _), _.updateTimestamp(_, _))
 
   /** @group Instances */
   implicit val ScalaBigDecimalMeta: Meta[BigDecimal] =
     BigDecimalMeta.imap(BigDecimal.apply)(_.bigDecimal)
 
+  import doobie.implicits.javasql.DateMeta
+
   /** @group Instances */
   implicit val JavaUtilDateMeta: Meta[java.util.Date] =
     DateMeta.imap[java.util.Date](a => a)(d => new java.sql.Date(d.getTime))
-
-  /** @group Instances */
-  implicit val JavaTimeInstantMeta: Meta[java.time.Instant] =
-    Basic.one[java.time.Instant](
-      Timestamp,
-      List(Char, VarChar, LongVarChar, Date, Time),
-      _.getObject(_, classOf[java.time.Instant]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaTimeLocalDateMeta: Meta[java.time.LocalDate] =
-    Basic.one[java.time.LocalDate](
-      Date,
-      List(Char, VarChar, LongVarChar, Timestamp),
-      _.getObject(_, classOf[java.time.LocalDate]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaLocalTimeMeta: Meta[java.time.LocalTime] =
-    Basic.one[java.time.LocalTime](
-      Time,
-      List(Char, VarChar, LongVarChar, Timestamp),
-      _.getObject(_, classOf[java.time.LocalTime]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaTimeLocalDateTimeMeta: Meta[java.time.LocalDateTime] =
-    Basic.one[java.time.LocalDateTime](
-      Timestamp,
-      List(Char, VarChar, LongVarChar, Date, Time),
-      _.getObject(_, classOf[java.time.LocalDateTime]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaOffsetTimeMeta: Meta[java.time.OffsetTime] =
-    Basic.one[java.time.OffsetTime](
-      TimeWithTimezone,
-      List(Char, VarChar, LongVarChar, Timestamp, Time),
-      _.getObject(_, classOf[java.time.OffsetTime]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaOffsetDateTimeMeta: Meta[java.time.OffsetDateTime] =
-    Basic.one[java.time.OffsetDateTime](
-      TimestampWithTimezone,
-      List(Char, VarChar, LongVarChar, Date, Time, Timestamp),
-      _.getObject(_, classOf[java.time.OffsetDateTime]), _.setObject(_, _), _.updateObject(_, _))
-
-  /** @group Instances */
-  implicit val JavaZonedDateTime: Meta[java.time.ZonedDateTime] =
-    Basic.one[java.time.ZonedDateTime](
-      TimestampWithTimezone,
-      List(Char, VarChar, LongVarChar, Date, Time, Timestamp),
-      _.getObject(_, classOf[java.time.ZonedDateTime]), _.setObject(_, _), _.updateObject(_, _))
 
 }
